Remove CodeBuild-specific build & deploy logic.
At this point I've spent more time than I want trying to get CodeBuild
and CodeCommit to play nice. The next step would probably be to look
into using CodePipeline to glue them together. Given how simple this
project is, it's not currently worth the extra time it would take.
Because of this, I've removed all of the CodeBuild-specific logic and
organization from the build and deploy scripts.

This also updates the README with information about the build & deploy
options, as well as information for setting up a development environment
from scratch.