api: JWT verifier doesn't create method-specific auth policies.
Previously the JWT verifier would respond with an auth policy that was
specific to the method being called. However, this prevented the
responses from being cached (calls to a different method within the
cache period would fail). Now we respond with a policy that allows
access to the entire API (for this environment) so that we can cache the
token responses.